requirements.txt
flask==3.1.0
flask-cors==5.0.1
azure-storage-blob==12.24.0
python-dotenv==1.0.1
gunicorn==23.0.0
